Added a builder script
The build script reduces the amount of IO on the disk when starting a new bash session.
This should speed up the startup time of the shell when under heavy IO load. There is
also a short reloader command that calls the builder script when things are modified to
avoid hot reloading of the scripts in the directory.